requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

Pillow==2.9.0
easy-thumbnails==2.5
html5lib==1.0.1

# django-social-auth is no longer maintained, and only has old-style migrations
# django-social-auth==0.7.28
# Use a fork to which we've added Django migrations, so we can keep using it a little
# longer.
git+https://github.com/caktus/django-social-auth@add_migrations#egg=django-social-auth

psycopg2==2.7.3.2
gunicorn==19.3.0
django-secure==1.0
raven==5.5.0

django-constance[database]==1.1
django-picklefield==1.0.0

django-multiselectfield==0.1.8

django-model-utils==2.3.1
django-redis-cache==1.6.0
redis==2.10.3
hiredis==0.2.0

graypy==0.2.9
